LayoutTests/platform/ios-simulator/editing/pasteboard/do-not-copy-unnecessary-styles-2-expected.txt

This test verifies that WebKit does not erroneously clone nodes hierarchy when copying text with background color.
"Hello" should be blue and has yellow background but you should not see red border.

You should not see any borders:
| <span>
|   style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|   "Hello<#selection-caret>"
